# openapi.json
{"openapi":"3.1.0","info":{"title":"简单学机器学习API","description":"为「简单学机器学习」APP提供的后端API","version":"1.0.0"},"paths":{"/api/v1/content/":{"get":{"tags":["content"],"summary":"Read Content","operationId":"read_content_api_v1_content__get","parameters":[{"name":"module","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"description":"模块: math/ml/dl","title":"Module"},"description":"模块: math/ml/dl"},{"name":"subcategory","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"description":"子分类","title":"Subcategory"},"description":"子分类"},{"name":"skip","in":"query","required":false,"schema":{"type":"integer","default":0,"title":"Skip"}},{"name":"limit","in":"query","required":false,"schema":{"type":"integer","default":100,"title":"Limit"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Content"},"title":"Response Read Content Api V1 Content  Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/content/{content_id}":{"get":{"tags":["content"],"summary":"Read Content By Id","operationId":"read_content_by_id_api_v1_content__content_id__get","parameters":[{"name":"content_id","in":"path","required":true,"schema":{"type":"integer","title":"Content Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Content"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/content/generate/":{"post":{"tags":["content"],"summary":"Generate Content","operationId":"generate_content_api_v1_content_generate__post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ContentCreate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Content"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/content/init_math":{"post":{"tags":["content"],"summary":"Init Math Content","description":"批量初始化数学内容，返回新增记录数","operationId":"init_math_content_api_v1_content_init_math_post","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/v1/content/init_ml":{"post":{"tags":["content"],"summary":"Init Ml Content","description":"批量初始化机器学习内容，返回新增记录数","operationId":"init_ml_content_api_v1_content_init_ml_post","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/v1/content/update/":{"post":{"tags":["content"],"summary":"Update Content","operationId":"update_content_api_v1_content_update__post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ContentUpdateRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/search/":{"get":{"tags":["search"],"summary":"Search Content","operationId":"search_content_api_v1_search__get","parameters":[{"name":"query","in":"query","required":true,"schema":{"type":"string","description":"搜索关键词","title":"Query"},"description":"搜索关键词"},{"name":"module","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"description":"模块过滤","title":"Module"},"description":"模块过滤"},{"name":"skip","in":"query","required":false,"schema":{"type":"integer","default":0,"title":"Skip"}},{"name":"limit","in":"query","required":false,"schema":{"type":"integer","default":10,"title":"Limit"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SearchResults"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/chart/{chart_id}":{"get":{"tags":["utils"],"summary":"Get Chart Image","operationId":"get_chart_image_api_v1_chart__chart_id__get","parameters":[{"name":"chart_id","in":"path","required":true,"schema":{"type":"string","title":"Chart Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/":{"get":{"summary":"Root","operationId":"root__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/health":{"get":{"summary":"Health Check","operationId":"health_check_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}}},"components":{"schemas":{"Content":{"properties":{"module":{"type":"string","title":"Module"},"subcategory":{"type":"string","title":"Subcategory"},"title":{"type":"string","title":"Title"},"content_body":{"type":"string","title":"Content Body"},"python_code":{"type":"string","title":"Python Code"},"formulas":{"anyOf":[{"type":"object"},{"type":"null"}],"title":"Formulas"},"charts_data":{"anyOf":[{"type":"object"},{"type":"null"}],"title":"Charts Data"},"tags":{"anyOf":[{"items":{"type":"string"},"type":"array"},{"type":"null"}],"title":"Tags"},"id":{"type":"integer","title":"Id"},"created_at":{"type":"string","format":"date-time","title":"Created At"},"updated_at":{"type":"string","format":"date-time","title":"Updated At"}},"type":"object","required":["module","subcategory","title","content_body","python_code","id","created_at","updated_at"],"title":"Content"},"ContentCreate":{"properties":{"module":{"type":"string","title":"Module"},"subcategory":{"type":"string","title":"Subcategory"},"title":{"type":"string","title":"Title"},"content_body":{"type":"string","title":"Content Body"},"python_code":{"type":"string","title":"Python Code"},"formulas":{"anyOf":[{"type":"object"},{"type":"null"}],"title":"Formulas"},"charts_data":{"anyOf":[{"type":"object"},{"type":"null"}],"title":"Charts Data"},"tags":{"anyOf":[{"items":{"type":"string"},"type":"array"},{"type":"null"}],"title":"Tags"}},"type":"object","required":["module","subcategory","title","content_body","python_code"],"title":"ContentCreate"},"ContentUpdateRequest":{"properties":{"modules":{"items":{"type":"string"},"type":"array","title":"Modules"},"subcategories":{"anyOf":[{"items":{"type":"string"},"type":"array"},{"type":"null"}],"title":"Subcategories"}},"type":"object","required":["modules"],"title":"ContentUpdateRequest"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"SearchResults":{"properties":{"results":{"items":{"$ref":"#/components/schemas/Content"},"type":"array","title":"Results"},"total_count":{"type":"integer","title":"Total Count"}},"type":"object","required":["results","total_count"],"title":"SearchResults"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}